refactor: [#59372] dependency inject sync_tool
The motivation for this commit was to add typing to lines alike:
```python
    sync_tool = context["user_context"]["sync_tool"]
```
As a means of improving editor support in the codebase, however instead
of simply adding typing alike this:
```python
    sync_tool: SyncTool = context["user_context"]["sync_tool"]
```
The proper route of converting the code to dependency injection was
taken. This change in itself was trivial, and should have been nothing
but a simple refactoring, however it was made severely more challenging
by the test-suite, which does not utilize dependency injection
whatsoever, instead favoring to do massive amounts of mocking all over
the place.

It was even noticed that some tests have interdependencies and that the
test-order is critical for the test-suite to function reliably, #60178
has been made to track this issue separately as it was deemed out of
scope to fix here, as this was meant to be nothing but a minor
refactoring.