fix!: [#59372] weird return condition in `format_converted_objects`
This commit attempts to fixes a weird return condition within the
`format_converted_objects` method, where a for-loop is started, but
always returns on the first iteration.

The code as it stands has a very unintuitive structure, and doing an
attempt at cleaning up the code, potentially by entirely eliminating
the for-loop, the idea that maybe the code has this structure
unintentionally was raised.

The weird return condition was introduced in !376 (3cd83578), and
appears to have been introduced by a mistake, which was later hidden
underneath a proper return statement that is never reached.

The pull request that introduced the change is approved, but with a
comment that the reviewer does not feel like they have the insight to
review the code, the author of the change states that they do not feel
confident in the change either, but that they feel like deploying it may
give insights to whether it is a good change or not.

The corresponding test-cases also have some hints that seem to indicate
that the current behavior is wrong, but what the correct and expected
behavior is here is very doubtful.

BREAKING CHANGE:

I legitimately have no idea if this change is correct or not, and all
attempts at figuring out whether it is or not has just left me more
confused, thus this change is made entirely based on a gut feeling, and
because of this, I have decided to release it as a breaking change.