refactor: [#59372] use first + unpacking instead of [0][0]
This commit uses a combination of `more_itertools.first` and tuple
unpacking to avoid a hard-to-read tuple unpacking follow by a list
indexing.

Using `first` allows us to communicate that we are specifically reading
the first element, thus blatantly ignoring all other matches, unlike if
`one` or `only` had been used to assert the number of elements, and
communicating that all matches had been handled.