fix: [#59372] fix bad CPR number typing
This commit fixes a badly typed GraphQL query that was introduced in
4.17.1 of the LDAP integration (!458 / cd5205d2). The issue was
introduced while rewriting the `find_mo_employee_uuid` method to using
the autogenerated graphql client.

The root-cause of the issue was lack of integration testing.
If we had, had integration tests, the issue would easily have been detected.