feat(amqp): [#59702] introduce LDAP AMQP
This commit introduces an AMQP queue for LDAP changes.

It is introduced via an entirely separate AMQP system from the MO system,
and currently listens only on a single routing key: `dn`.

Events are still generated via the same faulty mechanism (see related
issues on #59702), however while this does not resolve the issue of
dropped LDAP changes, it helps to guard against them.

The integration now takes a new environmental variable: `LDAP_AMQP__URL`,
if set the ldap AMQPSystem will be operated on this AMQP instance.
If not provided, it will default to the value set in `FASTRAMQPI__AMQP__URL`.