chore(jinja)!: [#59372] remove the parse_datetime jinja filter
This commit removes the `filter_parse_datetime` and the corresponding
`parse_datetime` jinja-filter currently only in use on Holstebro
development.

The function is the only thing keeping our dependency on pandas alive,
specifically to call the heuristic `pandas.to_datetime` function which is
used to try to guess a datetime without specifying the datetime format.

The function entered the codebase in !35 (27e2876d) originally to support
Alleroed's field mapping, later being modified (b168a598) to support out
of bounds dates, with a year check against 2000, due to pandas' internal
handling of dates, see:

* https://pandas.pydata.org/pandas-docs/stable/user_guide/timeseries.html

Find the section on "Timestamp limitations"

The code was further modified in !65 (1a58a6c3) and !93 (991a4357),
before being moved to where it is now in !262 (91fd9590).

While the code was introduced to support Alleroed, nowadays Alleroed does
not in fact utilize the function, instead favoring to use `strptime` with
a specified format string.

This solution seems vastly favorable to guessing the format, as we care
greatly about the correctness of data, thus Holstebro development should
take a similar approach to parsing dates.

This commit also adds `pytz` as a development dependency as it is used
for a single unit-test. It was previously pulled in by pandas as a
side-effect, in the future we should consider if we can use the standard
`zoneinfo` standard-library instead, see: #60295.

BREAKING CHANGE:

The `filter_parse_datetime` function and the corresponding `parse_datetime`
jinja-filter has been removed, datetimes should rather be parsed using
`strptime` if needed.