feat: [#59732] reimplement listen_to_changes_in_mo
This commit reimplements the functionality underneath the
`listen_to_changes_in_mo` configuration variable.

Prior to this change, the LDAP integration would actually listen to
changes in MO even when this flag was set to false, however all messages
that would be received would be rejected instantly.

With this change the LDAP integration does it fact not listen to changes
in MO, simply by not binding a router to the AMQPSystem, and thus not
getting any messages whatsoever.