feat!: [#57671] introduce service_api role
This commit introduces the `service_api` role, which controls whether
OS2mo's ServiceAPI can be accessed or not. The goal of this change is to
restrict access to the ServiceAPI until it can be completely eliminated.

Attempting to access the ServiceAPI without the proper permission now
yields an authorization error stating that the ServiceAPI is gone.

BREAKING CHANGE:

After this commit only clients with the `service_api` role can access
the ServiceAPI.