feat(graphql): [#59178] add engagement filter to org_units
This commit adds the option to filter org_units by engagements.

The implementation works by first resolving the engagement filter to a
list of engagement UUIDs, which is then used to fetch the engagement
objects, from which a list of org_unit UUIDs is extracted, finally
this list of UUIDs is used to filter the UUID query.