feat: [#60336] introduce read-only flag
The commit adds the `LDAP_READ_ONLY` environmental variable, which
allows the integration to be configured into a read-only mode, to ensure
that the LDAP server's content is not modified by the integration.

Note that this flag does not seem to actually work currently, see:

* https://github.com/cannatag/ldap3/issues/1008

However we introduce it and configure it anyway on the chance that it
may work in the future, and as an attempt to declare our intent in
different server configurations.