feat(db): [#60000] set isolation_level to serializable
This commit sets the default transaction isolation to serializable,
prior to this change we were running with the default of read committed,
however the code did not seem to take this into account.

Following a longer discussion regarding the implementation of ETags, it
was decided that we want to run all of OS2mo with the serializable
isolation level as this ensures maximum correctness, potentially at the
cost of performance though.