fix: [#59372] ensure cleanup emits all cleanup events
This commit fixes an issue in the LDAP cleanup function, where refresh
events would not be published as expected.

The issue was a faulty loop-condition in which the list of UUIDs to
refresh was potentially overwritten at each iteration of the loop, thus
only keeping the result of the last iteration, rather than collection
UUIDs throughout all iterations.