fix: [#60337] deprecate *_refresh mutator queue
It never worked properly; to send to a specific queue, a message was
sent to the "" exchange with the queue name as routing key. This routing
key persisted all the way to the integration, which would receive an
unexpected routing key.

We will introduce an `exchange` parameter instead. This also decouples
the refresh caller from the implementation details of the integration's
queue names.