refactor: [#60478] convert import_class to factory pattern
This commit removes the dynamic nature of `import_class` and replaces it
with a factory pattern instead. This means that `import_class` can no
longer be used to load arbitrary code from the standard library or any
other libraries available in the python environment.

The new implementation rather has a specified list of inputs and their
corresponding output values, such that only a curated selection of
classes can be utilized.

Thus in the future to utilize models outside of the curated list a
code change must be made such that the desired model enters the curated
list.