refactor: [#59851] refactor create to per object creation
This commit splits the `create` function on `DataLoader` into two
function, an orchestrator which starts a number of async tasks, and a
helper which can upload a single object.

This is purely a refactoring, and thus no test changes should have been
needed, however the tests are incorrectly mocking the behavior of the
underlying `legacy_model_client` and as such had to be modified during
the refactoring. In theory it should be possible to have fixed the tests
before this refactoring, however as the ultimate goal is to replace the
`legacy_model_client` with our autogenerated GraphQL client this seems
like a waste of work, and thus have not been done.

Additionally this commit also adds a test for exception handling during
creates in MO, which was not test covered before.