feat: [#59851] refactor create_object to per type creation
This commit splits the `create_object` function on `DataLoader` into
three functions, an orchestrator which checks the type of its argument
and calls the corresponding type-specific create helper to create an
entity of that type.

This is purely a refactoring, and thus no test changes should have been
needed, however the tests are incorrectly mocking the behavior of the
underlying `legacy_model_client` and as such had to be modified during
the refactoring. In theory it should be possible to have fixed the tests
before this refactoring, however as the ultimate goal is to replace the
`legacy_model_client` with our autogenerated GraphQL client this seems
like a waste of work, and thus have not been done.

Additionally this commit also adds some tests that are needed after the
refactoring, as the code was not properly tested before, merely tested
to the point that it had code coverage.