refactor: [#59372] remove check_dar_scope
This commit removes the mapping validator that blocks mapping DAR
addresses from LDAP to MO, thus relaxing validations on the
configuration.

The block was documented in the code via a comment, arguing its
existence by the following rationales:

1. DAR is not available in Greenland

The counter argument for this, is that the integration runs in Denmark
as well, where DAR is is indeed available.

2. DAR UUIDs are not present in LDAP

The counter argument for this, is that they may very well be, we do not
know which fields are mapped in future customers LDAP instances nor do
we know which fields our current customers may introduce in the future.

3. DAR UUIDs may not have the right format in LDAP

The counter argument for this, is that it applies to literally all
fields that we read from LDAP, which is exactly why we have
jinja-templates for the mapping.

So if the DAR UUIDs in LDAP do not have the right format, we will simply
translate them via a jinja-filter until they do, just as we are doing
with the ADGUID already.