refactor: [#59372] remove check_get_uuid_functions
This commit removes the mapping validator that checks if `get_*_uuid`
functions are being utilized correctly (always called with user-keys).
Thus relaxing validations on the configuration.

The rationale for the removal is as follows;

1. The validator checks methods on the converter, not the template context

The code being validated did not necessarily equate the code that was
available to configurations, since not every method on the converter was
exposed in the template context, this was especially true after a lot of
the recent refactorings has moved functions out of the converter itself.

This could be remedied by reconfiguring the checker to run on functions
in the template context / template environment.

2. The validator attempts to do static analysis, but does it poorly

The current checking implementation attempts to do static analysis
simply by processing the strings, searching for the function name and
looking forward to a corresponding parenthesis pair.

This does not work in the general case, as arguments to function can be
other function calls, thus having parenthesis themselves, or arguments
could be hard-coded strings that contain parenthesis.

The core issue core, is that we are trying to parse jinja-templates
using regular language constructs, whereas a proper parser would have to
be recursive, such as a context-free grammar.

This could be remedied by reimplementing the entire validator on the
parsed jinja-templates utilizing the parsing available in the jinja2
library.

3. The validator utilizes cached information

The implementation makes heavy use of the info-dicts cache to do its
checking, thus the data we validate against may be out-of-date, thus
rejecting legal configurations, although this is unlikely since the
cache is created immediately before the check.