feat: [#60126] make load_ldap_cpr_object return multiple accounts
This commit ensures that the `load_ldap_cpr_object` finds all accounts
with a given CPR number instead of merely finding one and failing on
duplicates.

This thus fixes an bad invariant in the code, namely assuming that only
one account would be found for each CPR number. We do know in fact that
multiple accounts does exist for the same CPR number, usually as a
result of multiple independent deployments, political titles, etc.