feat(graphql): [#59681] calculate primary based on all the users engagements
Earlier the field is_primary was found by checking whether the associated primary class
was considered primary or not by chehcking if its scope was 3000 or more. This lead to a
problem with engagements with a primary scope of 1000 ("Ansat - ikke i løn"). Even if this
was the only engagement it would not be considered primary.

Now is_primary is calculated by checking each engagement against the users other engagements.
This allows engagements with "Ansat - ikke i løn" to be primary if (and only if) it is the
only engagement.