feat(graphql): [#60575] support disabling old GraphQL versions
This commit introduces a new environmental variable `MIN_GRAPHQL_VERSION`
which can be used to set the lowest GraphQL version to expose.

It currently defaults to 2, the lowest GraphQL version available, in
order to be backwards compatible, but the goal for the future is to set
it to the newest GraphQL version available, such that only the newest
version is exposed by default.