test(graphql): [#58204] fix test_parent_changes
This commit fixes the `test_parent_changes` to actually function on real
data, the test was introduced in a broken state due to a misassumption
around how the `parent_uuid` and `parent` fields works.

As it turns out `parent_uuid` will return the UUID of the root
organisation rather than `None` when an organisation unit is a root in
the organisation unit tree, while the `parent` lookup will return `None`.