feat(graphql): [#59178] introduce temporality selectors
This commit introduces selectors to the GraphQL response temporality
layer allowing API users to not only specify the filter time, but also
the time for which data is returned.

Specifically it is now possible to set the `at` time for current data,
such as with the following query:
```graphql
query ReadEmployeeAt2010($uuid: UUID!) {
  employees(filter: {uuids: [$uuid]}) {
    objects {
      current(at: "2010-01-01") {
        validity {
          from
          to
        }
        name
      }
    }
  }
}
```
Which will return the data that was active on the 1st of January 2010,
such as:
```json
{
  "data": {
    "employees": {
      "objects": [
        {
          "current": {
            "validity": {
              "from": "2000-01-02T00:00:00+01:00",
              "to": "2019-12-31T00:00:00+01:00"
            },
            "name": "John Deere"
          }
        }
      ]
    }
  }
}
```
Similarly validities can also be queried with an interval, given `start`
and `end` timestamps:
```graphql
query ReadEmployeeDataBetween2000And2010($uuid: UUID!) {
  employees(filter: {uuids: [$uuid]}) {
    objects {
      validities(start: "2000-01-01", end: "2010-01-01") {
        validity {
          from
          to
        }
        name
      }
    }
  }
}
```
Which will return the data that was active in the interval between the
1st of January 2000 and the 1st of January 2010, such as:
```json
{
  "data": {
    "employees": {
      "objects": [
        {
          "validities": [
            {
              "validity": {
                "from": "1900-01-01T00:00:00+01:00",
                "to": "2000-01-01T00:00:00+01:00"
              },
              "name": "Beforaza Deere"
            },
            {
              "validity": {
                "from": "2000-01-02T00:00:00+01:00",
                "to": "2019-12-31T00:00:00+01:00"
              },
              "name": "John Deere"
            }
          ]
        }
      ]
    }
  }
}
```
The introduction of this new behavior is fully backwards compatible with
the old behavior, as it is only activated when `at` and `start` / `end`
respectively is included on the temporal axis, something that was
illegal in previous versions of the schema.

The behavior of `start` and `end` is identical to the uses in the
filters, such that an unset `start` is interpreted as 'now', etc.

If `at` or `start` / `end` is specified the new behavior takes
precedence over the old behavior of inheriting the result time from the
filter time.

The functionality is backwards compatible to GraphQL v21, as this is
when date handling was corrected, however the new syntax is accepted all
the way back to GraphQL v14 although with the old date handling behavior.