feat(graphql): [#59178] disable DataLoader caches
This commit disables caches on our GraphQL DataLoaders as they only use
UUID as cache-key, but really depend on a much larger context for
caching and thus exhibit incorrect cache characteristics.

The appropriate for this issue is to ensure that arguments are passed
through the stack, and thus through the dataloader, rather than around
the stack, as-is the case right now with GraphQL dates, that pass around
the stack via the use of starlette_context.

Thus this is simply a temporary removal of the cache to eliminate a
potential footgun, potentially at the cost of performence. If it turns
out that the performance impact of this change is too great, it can be
reverted at the cost of reintroducing the footgun.