refactor: [#59178] added missing typing inside function
This commit adds missing typing on two contextual objects, whose lack of
typing almost lead to a cleanup being merged, which would have broken
the integration.

The core issue is not actually the lack of typing, but the lack of
testing, but this is the simplest way to ensure that the code is not
removed too soon, as doing so would break the typing.