feat(config)!: [#60126] validate apply_discriminator settings
This commit adds a layer of validation to the discriminator settings,
specifically it is now enforced that `DISCRIMINATOR_FUNCTION` and
`DISCRIMINATOR_VALUES` must be set if `DISCRIMINATOR_FIELD` is set.

BREAKING CHANGE:

This is a breaking change as some settings that were optional before are
now required, however this should not be a problem for any deployments as
a configuration without having all three environmental variables set does
not make a lot of sense, yet is it still a breaking change.