refactor: [#60126] convert DN functions to return sets
This commit refactors a couple of Distinguished Name functions to return
sets of DNs instead of lists, in order to communicate that we literally
have no idea about the order of the results, and that we should not
depend on the order of the results.

Additionally it adds a few TODOs and improves typing in a few places.