feat: [#60126] introduce first_included
This commit introduces `first_included` a function to apply our
discriminator to a set of DNs in order to decide which LDAP account is
the primary.

The code is inspired by function of the same name in the old LDAP
integration, and aims to implement it in the structure of the new
integration.

The method really should be called `apply_discriminator`, but a prior
attempt at implementing the discriminator logic has this name, an
attempt which unfortunately failed at implementing equivalent behavior
with the old LDAP integration.