feat: [#60126] add attributes parameter to get_ldap_object
This commit adds an `attributes` parameter to the `get_ldap_object`
function, such that the caller can decide which attributes to fetch.

The function retains its old behavior of fetching *ALL* attributes, when
not explicitly given a list of attributes to fetch, although this
behavior may change in the future, such that it is required to define
which attributes should be fetched to avoid overfetching.