feat(graphql): [#60126] introduce engagement is_primary filter
This commit introduces the `is_primary` filter to `engagements` in
GraphQL to allow filtering engagements on whether they are primary or
not.

The current implementation of the filter is sub-optimal, but we prefer
to establish the interface early, such that integration can use the
filter and thereby avoid technical debt from implementing the filter in
client code.

In the future the filter can then be implemented in a more performant
way, without breaking any integrations, instead only improving their
performance over time.

The effect of using the filter is identical to reading `current.is_primary`
with `at=now()` and filtering on the returned value in client code.