feat(amqp): [#60890] add request_id to AMQP messages
This commit adds a `request_id` to all AMQP message handling logs.

The goal of the `request_id` is that each handler call can be
distinguished from each other when inspecting the log stream.

Thus to pinpoint a problem, a problematic log message can be found, and
a filter can be set for either the `message_id` to simply see all logs
regarding that message (including retries) or by `request_id` to only
see the specific log trace containing the log message in question.

This is especially useful with drawn out message processing in which it
is hard to distinguish delays between log lines from sleep_on_error
delays.