feat: [#60126] reimplement process_address handler
This commit reimplements the `process_address` handler, such that it now
merely translates an address uuid into the corresponding employee or
org-unit uuid and then calls the appropriate downstream handler.

Prior to this change this call was done using the `unpack_payload`
function, which was highly generic, whereas now it is done using an
address specific query.