fix: [#61047] handle empty discriminator input set
This commit ensures that the `first_included` function can handle the
empty input set, returning `None` whenever it receives such a set.

Prior to this change, the function would fail with an exception when the
empty input set was given, however it makes more sense to simply return
`None` to indicate that no good account could be found.