feat(graphql): [#60126] bulk refresh employees
This commit changes our `employee_refresh` mutator to take a list of
UUIDs instead of a single UUID, such that an `asyncio.gather` can be
eliminated, and multiple calls to OS2mo can be combined to one.

Single UUID calls to `employee_refresh` now call with a single element
list.