feat: [#60126] run sync ldap search in async context
This commit removes an sync-to-async wrapper using `loop.run_in_executor`
in favor of simply running sync code in our async handler.

This is very, very bad as it blocks the entire event-loop, but it aligns
this piece of the code with the rest of the project, where we also run
sync code in our async handlers, thus it does not actually worsen the
quality of the code by a lot.

The change does however ensure that we do not utilize our non thread
safe LDAP connector from multiple threads, which should hopefully
reduce the number of issues arising from invoking undefined behavior.

In the future all of these sync calls in our async handlers should be
refactored, such that we instead do async calls.