fix: [#60126] fix the Alleroed username generator
This commit fixes an issue with the Alleroed username generator, which
was introduced by a supposed refactoring in 10.1.0 (759ed124) of the
integration.

The issue was introduced to the code that ensures that no previously
generated username can ever be reused in the future.
The issue did not cause the generation of previously generated
usernames, but rather simply made the code crash whenever it was
executed.

The refactoring that introduced the issue, removed a function which was
deemed to be unused, however due to poor typing, the use from within the
username generator was not found.

This commit reintroduces the removed functionality, and introduces
proper typing to the username generator, such that a similar issue
cannot be introduced in a future version, as it will now be detected by
the static type checking.

Additionally, the lack of integration testing also kept us from finding
the issue, had the integration been properly integration tested, the
issue would have been trivial to find, but unfortunately no such tests
exist for the integration.