feat(ldap): [#60126] async ldap connection
This commit switches the LDAP connection strategy from `RESTARTABLE` to
`ASYNC`, thereby changing how we communicate with the LDAP server.

With `RESTARTABLE` the connection is synchronous / blocking and non
thread-safe, thus before this commit any LDAP query effectively stalls
the entire integration while waiting for a reply, additionally running
queries in parallel results in undefined behavior, which may explain why
the integration has had a tendency to randomly freeze up.

With `ASYNC` the connection is asynchronous / non-blocking and
thread-safe, thus after this commit LDAP queries do not stall the
program, and a higher level of parallelization can be achieved,
additionally running queries in parallel is not an issue.

The `ASYNC` strategy is configured with a configurable timeout, such that
if queries take longer than the timeout they will time-out and an
exception will be raised, which will requeue and retry the message.
Similarly a receive timeout is in place, which will terminate the entire
LDAP connection if it goes stale for more than another configurable time.

This change has been made possible my the many, many refactorings in the
previous merge requests, and we hope it will ensure a more smooth
operation of the integration with more predictable performance
characteristics.