feat: [#59367] allow mapping to msDS-cloudExtensionAttributes
This commit adds `msDS-cloudExtensionAttributes`s to the allowlist for
mapping attributes that do not need to be found in the LDAP schema
before running the integration.

It is added merely as a counterpart to the existing `extensionAttribute`s,
without any understand of why these must be allowed outside of the
schema, why they are not in the schema or similar.