feat: [#59367] raise all check-attribute errors at once
This commit ensures that all invalid attributes errors are thrown on the
first run of the integration, as opposed to only throwing the first
occurring error.

By throwing all the errors at once the operator of the integration can
check and fix all the configuration issues at once, instead of fixing
one, then rerunning and redeploying the integration just to find that
another issue is present in the configuration.