fix(ldap): [#61143] fix ldap_compares behavior when attributes are not found
This commit fixes an issue with `ldap_compare` function, which was
introduced  during the transition to asynchronous LDAP connectivity in
version 10.6.0 (b4e3e50e) of the integration.

The synchronous implementation of the `ldap_compare` function would
return `False` when an attribute was not found in LDAP, as documented
here: https://ldap3.readthedocs.io/en/latest/compare.html

Whereas the current implementation raises an "Unknown comparison result"
`ValueError` when this occurs, thus exhibiting wildly inequivalent
behavior.

This commit attempts to ratify this by bringing the asynchronous
implementation in line with the prior synchronous implementation.

Additionally a handful of tests have been added for the `ldap_compare`
function.