fix: [#61142] handle temporally separated primary engagements
This commit reworks the logic within `get_primary_engagement` such that
it handles multiple temporally spread primary engagements correctly.

Prior to this change, the integration would complain about multiple
primary engagements, even if they were non-overlapping and we could
simply pick the currently active one.

Now the integration only complains if multiple current primary
engagements overlap. The implementation is not perfect by far, but it is
much closer to the desirable behavior than the prior implementation,
bringing it to a point where it is no longer the biggest challenge with
the integration.

In the future this code should be revisited, but for now this
implementation is good enough. The future implementation should probably
delegate a lot of the work to OS2mo itself, by the use of an `is_primary`
GraphQL filter.