refactor: [#60126] spread run_discriminator
This commit spreads the `run_discriminator` parameter from the
`single_object_search` function to `get_ldap_dn`, `get_ldap_object` and
the callers of these functions.

This commit is a pure refactoring to move the argument further up the
call-stack where reasonable arguments can be given that the
discriminator does not actually need to be called.