feat: [#61188] introduce LDAP UUID handler
This commit introduced an LDAP `UUID` handler, which functions similarly
to the old LDAP DN handler, by calling the `import_single_user` handler
after converting the UUID to a DN.

The old LDAP DN handler has been converted to simply emit LDAP `UUID`
events, and is planned to be removed once event emitters are all
emitting `UUID`s instead of DNs. That change will be coming in a future
commit, this simply lays the groundwork for it.