feat(ldap,amqp): [#61188] no DNs publishing
This commit removes the `publish_dns` function and thus ensures that we
no longer publish DNs to the event system, as these may have changed
when the message actually gets processed.

We now instead only publish UUIDs (objectGUID / entryUUID) as these are
not subject to change when the object gets renamed or moved, thus
ensuring that the event system works as expected.

The `process_dn` event handler persists for a bit to ensure that all
AMQP messages left in this queue gets converted to UUIDs before its
removal in a future change.