feat(ldap): [#59367] support templates in apply_discriminator
This commit adds a new `discriminator_function` to the integration;
`"template"` which takes jinja templates as `discriminator_values` and
evalutes the template with the `dn` and retrieved `discriminator_field`
value as context.

The primary DN is decided by when this template evalutes to the string
`"True"` and otherwise works similar to how the `"include"`
`discriminator_function` works.