feat(api, mo2ldap): [#61146] introduce template test endpoint
This commit introduces a new endpoint `/Inspect/mo2ldap/{uuid}`, which
allows the caller to see how the MO employee at with `uuid` would be
mapped into LDAP if they were being synchronized.

The output of the new endpoint is a dictionary with the keys being the
`json_key`s from the configuration, and the values being two element
tuples, the first element being the mapped values to be written into
LDAP and the second element being whether the values should be deleted
in LDAP or not.

To introduce the endpoint a dry-run flag has to be added to the
`listen_to_changes_in_employees` handler, such that the synchronization
code could be run in a safe mode without ever writing to LDAP.

Additionally the logic within the `listen_to_changes_in_employees`
handler had to be adjusted such that determining what to write, and the
actual writing has been separated, this restructuring is supposed to be
a pure refactoring.