feat: [#59181] support mo2ldap on openldap
This commit finalizes the changes required for the mo2ldap part of the
development environment to be function, such that the integration can
finally be run locally during development.

The changes are quite extensive in order to support both the standard
LDAP protocol utilized by OpenLDAP and the Active Directive LDAP dialect
used by Microsoft.