fix(ldap2mo): [#59367] only lookup CPR accounts if we have CPR
This commit fixes an issue where we would attempt to lookup LDAP
accounts based on CPR numbers even when no such CPR number was found.

After this change we simply use the event-generated DN to create the
user if no CPR number is found on the account, and thus we cannot
find other potential accounts.

In the future we may want to change this behavior to simply reject
accounts without CPR number set, but for now this can be done using the
template discriminator introduced in version 10.13.0 of the integration.