feat(mo2ldap): [#61315] emit new messages on requeue
This commit adds a horrendous hack to work around the broken state of
quorum queues in RabbitMQ and the fact that messages are not returned to
the back of the queue when requeued, but rather to the front of the queue.

To work around this issue, this commit now just replies positively to
all RetryExceptions after first emitting a new AMQP message with the
same UUID (ensuring it does in fact go the back of the queue).

This is very much a hack and by means encouraged to be replicated
anywhere. The real solution is replacing RabbitMQs broken system with
something better.