feat(api): [#61146] introduce MO UUID to LDAP DN endpoint
This commit introduces a new API endpoint: `/Inspect/mo/uuid2dn/{uuid}`
which takes a MO UUID and returns a set of corresponding LDAP DNs as
mapped via CPR number or IT system accounts.

This endpoint is useful for testing the account mapping when debugging
production issues, but is also quite useful for integration-testing
involving generating DNs using the `UserNameGenerator`.