fix(ldap2mo): [#61058] remove exception from load_mo_x_addresses
This commit fixes an issue with the `format_converted_objects` method,
specifically the part regarding address processing. This function is
responsible for finding existing addresses and figuring out how to
update them, or whether to create them / delete them instead.

The issue seems to arises when the code cannot find any existing addresses
in MO (i.e. the operation is a create), when this situation occurs the
`load_mo_employee_addresses` / `load_mo_org_unit_addresses` raises a
`NoObjectsReturnedException` instead of simply returning an empty list.
This exception then stops the processing.

It is difficult to pin down exactly when the issue was introduced, but
it seems to have been present for a very, very long time.

This commit fixes the issue by refactoring the `load_mo_x_address`
functions, such that they no longer raise an exception whenever no
elements are found, but rather simply return an empty list.

A similar change was made for `load_mo_employee_it_users` in version
10.18.0 of the integration (!689, 8144b39c).

The issue has not resulted in any erroneous writes, thus data integrity
should be entirely unaffected. The change may however have resulted in
blocked synchronizations, such that some entities in LDAP have not been
correctly updated. This may be problematic however as the errors have
simply skipped the problematic entries, and have not retried the AMQP
messages, thus no backlog of AMQP messages exist and in theory all LDAP
accounts will have to be resynchronized to resolve the issue for all
users.

It is unfortunate that we do not have integration tests covering this
scenario, but given that we now have a framework to develop these,
perhaps a regression integration test will be added for this in the
future. For now howver this commit simply resolved the issue and does
not in fact introduce a new integration test to avoid regressions in the
future.