fix(mo2ldap): [#61431] fix a bug in the Alleroed username generator
This commit fixes an issue with the `AlleroedUserNameGenerator`, the
generator implements a specific methodology unique to Alleroed.

The issue is in the code that converts the configured ITSystem user-key
to an UUID to lookup the ITUsers under the ITSystem.
Currently it crashes with an `AttributeError` error that a certain method
does not exist.

A similar issue was found with the user export checks, and corrected in
!702 (281d9a97), however this time for a different part of the code.

The issue was introduced in version 10.20.0 of the integration, with !693
(7bc34c9b), which reimplemented the `get_it_system_uuid` method as a
function, and in the process isolated it from the `LdapConverter` object.

Unfortunately during this process it was missed that the
`AlleroedUserNameGenerator` actually called the method directly on the
`LdapConverter` object.

This commit fixes the issue by refactoring the call to the new function,
replacing the old method call as it was done for most call-sites during
the reimplementation in the above mentioned pull request.

The issue has not resulted in any erroneous writes, thus data integrity
should be entirely unaffected. The change may however have resulted in
blocked synchronizations, such that some entities in LDAP have not been
correctly updated. This should have been a non-issue as the AMQP messages
should have been piling up, both serving to alert us to the issue, and
also ensuring that the backlog of broken synchronizations could be
rerun.

It is unfortunate that our static analysis typing tool mypy was not
unable to detect this discrepancy, seeing as the called method no longer
exists on the `LdapConverter` object, however due to lack of typing and
dynamically typed objects (specifically our context object) this was not
possible. This commit does however add the much needed type annotations,
and additionally also add a missing `UUID` cast from the original code.

The issue was a subtle one that was fairly not caught in the review
process, however it should have (and likely would have) been caught if
we had had integration tests. This commit does not introduce integration
tests, but a future commit will.