fix(ldap2mo): [#60256] fix the get_number_of_entries method
This commit fixes an issue with the `get_number_of_entries` method, the
method determines the number of MO objects to create for each LDAP
object found based on the cardinality of the data fields in LDAP.

The issue with the code was in handling the special case where the only
list value was an empty list or a set of empty lists.

The issue was introduced in version 4.13.0 of the integration with !443
(ad9375c5), which rewrote the `get_number_of_entries` method to make it
clearer to understand, but had the unforeseen side-effect of missing the
mentioned edge-case.

This commit fixes the issue by rewriting the method yet again, this time
with a deeper understanding of the desired functionality of the method.

The issue has not resulted in any erroneous writes, thus data integrity
should be entirely unaffected. The change may however have resulted in
blocked synchronizations, such that some entities in MO have not been
correctly updated. This may be problematic however as the messages have
simply as not needing synchronizations, and thus no backlog of AMQP
messages exist and in theory all LDAP accounts will have to be
resynchronized to resolve the issue for all users.

The issue was a subtle one that was fairly not caught in the review
process, however it should have (and likely would have) been caught if
we had had integration tests or unit-tests covering the corner cases.
While commit does not introduce integration tests, the prior commit did
add parametric unit-tests covering the problematic corner case.