feat(ldap2mo): [#60256] rework empty values handling in from_ldap
This commit reworks how empty values are handled within the `from_ldap`
method. Prior to this change empty values were being filtered
immediately upon template output, whereas now the template output is
saved as-is.

This change was made to support making two separate checks on the
template output, namely one check which ensures that all the required
fields are mapped and raises an exception if this is not the case, and
one check which ensures that the required fields actually have non-falsy
values and skips the entry if this is not the case.